# docdoku-server/docdoku-server-web/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v2.5",
    "title" : "DocdokuPLM API"
  },
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/workspaces/{workspaceId}/folders" : {
      "get" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Get root folders",
        "description" : "",
        "operationId" : "getRootFolders",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "configSpec",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FolderDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Create root folder",
        "description" : "",
        "operationId" : "createRootFolder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FolderDTO"
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{completePath}/folders" : {
      "get" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Get sub folders",
        "description" : "",
        "operationId" : "getSubFolders",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "completePath",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "configSpec",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FolderDTO"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{folderId}" : {
      "put" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Rename a folder",
        "description" : "",
        "operationId" : "renameFolder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FolderDTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Delete root folder",
        "description" : "",
        "operationId" : "deleteRootFolder",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{folderId}/documents" : {
      "get" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Get documents from given params",
        "description" : "",
        "operationId" : "getDocuments",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tagId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "query",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "assignedUserLogin",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "checkoutUser",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "filter",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "configSpec",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentRevisionDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Create document",
        "description" : "",
        "operationId" : "createDocumentMasterInFolder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "configSpec",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/folders/{folderId}/move" : {
      "put" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Move a folder",
        "description" : "",
        "operationId" : "moveFolder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FolderDTO"
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{parentFolderPath}/folders" : {
      "post" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Create a sub folder",
        "description" : "",
        "operationId" : "createSubFolder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "parentFolderPath",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FolderDTO"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ACLDTO" : {
      "type" : "object",
      "properties" : {
        "userEntries" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string",
            "enum" : [ "FORBIDDEN", "READ_ONLY", "FULL_ACCESS" ]
          }
        },
        "groupEntries" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string",
            "enum" : [ "FORBIDDEN", "READ_ONLY", "FULL_ACCESS" ]
          }
        }
      }
    },
...
    "ActivityDTO" : {
      "type" : "object",
      "properties" : {
        "step" : {
          "type" : "integer",
          "format" : "int32"
        },
        "relaunchStep" : {
          "type" : "integer",
          "format" : "int32"
        },
        "tasks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TaskDTO"
          }
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string",
          "enum" : [ "SERIAL", "PARALLEL" ]
        },
        "tasksToComplete" : {
          "type" : "integer",
          "format" : "int32"
        },
        "complete" : {
          "type" : "boolean",
          "default" : false
        },
        "stopped" : {
          "type" : "boolean",
          "default" : false
        },
        "inProgress" : {
          "type" : "boolean",
          "default" : false
        },
        "toDo" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "DocumentIterationDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "documentMasterId" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "iteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modificationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "checkInDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "title" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "revisionNote" : {
          "type" : "string"
        },
        "attachedFiles" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "instanceAttributes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InstanceAttributeDTO"
          }
        },
        "linkedDocuments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DocumentRevisionDTO"
          }
        },
        "documentRevision" : {
          "$ref" : "#/definitions/DocumentRevisionDTO"
        }
      }
    },
...
    "DocumentRevisionDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "documentMasterId" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "commentLink" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "checkOutUser" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "checkOutDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "description" : {
          "type" : "string"
        },
        "iterationSubscription" : {
          "type" : "boolean",
          "default" : false
        },
        "stateSubscription" : {
          "type" : "boolean",
          "default" : false
        },
        "documentIterations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DocumentIterationDTO"
          }
        },
        "workflow" : {
          "$ref" : "#/definitions/WorkflowDTO"
        },
        "path" : {
          "type" : "string"
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "publicShared" : {
          "type" : "boolean",
          "default" : false
        },
        "attributesLocked" : {
          "type" : "boolean",
          "default" : false
        },
        "acl" : {
          "$ref" : "#/definitions/ACLDTO"
        },
        "lastIteration" : {
          "$ref" : "#/definitions/DocumentIterationDTO"
        }
      }
    },
...
    "FolderDTO" : {
      "type" : "object",
      "properties" : {
        "path" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "home" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "InstanceAttributeDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "mandatory" : {
          "type" : "boolean",
          "default" : false
        },
        "locked" : {
          "type" : "boolean",
          "default" : false
        },
        "type" : {
          "type" : "string",
          "enum" : [ "TEXT", "NUMBER", "DATE", "BOOLEAN", "URL", "LOV" ]
        },
        "value" : {
          "type" : "string"
        },
        "lovName" : {
          "type" : "string"
        },
        "items" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/NameValuePairDTO"
          }
        }
      }
    },
...
    "NameValuePairDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        }
      }
    },
...
    "TaskDTO" : {
      "type" : "object",
      "properties" : {
        "closureComment" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "instructions" : {
          "type" : "string"
        },
        "targetIteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "closureDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "signature" : {
          "type" : "string"
        },
        "worker" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "NOT_STARTED", "IN_PROGRESS", "APPROVED", "REJECTED", "NOT_TO_BE_DONE" ]
        }
      }
    },
...
    "UserDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "login" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        },
        "membership" : {
          "type" : "string",
          "enum" : [ "READ_ONLY", "FULL_ACCESS" ]
        }
      }
    },
...
    "WorkflowDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "finalLifeCycleState" : {
          "type" : "string"
        },
        "activities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ActivityDTO"
          }
        },
        "abortedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "currentActivity" : {
          "$ref" : "#/definitions/ActivityDTO"
        },
        "currentStep" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
  }
}